Moved Travertino tests back; altered tox and CI
