+++ .github/workflows/ci.yml
@@ -59,6 +59,7 @@ jobs:
         - "gtk"
         - "iOS"
         - "toga"
+        - "travertino"
         - "textual"
         - "web"
         - "winforms"
@@ -67,8 +68,8 @@ jobs:
       build-subdirectory: ${{ matrix.subdir }}
       attest: ${{ inputs.attest-package }}
 
-  core:
-    name: Test core
+  test-core-and-travertino:
+    name: Test ${{ matrix.package }}
     runs-on: ${{ matrix.platform }}
     needs: [ pre-commit, towncrier, package ]
     continue-on-error: ${{ matrix.experimental }}
@@ -77,8 +78,13 @@ jobs:
       matrix:
         platform: [ "macos-latest", "ubuntu-latest", "windows-latest" ]
         python-version: [ "3.9", "3.10", "3.11", "3.12", "3.13" ]
+        package: ["core", "travertino"]
         include:
         - experimental: false
+        - package: "core"
+          env_suffix: ""
+        - package: "travertino"
+          env_suffix: "-trav"
 
     steps:
     - name: Checkout
@@ -110,16 +116,16 @@ jobs:
       run: |
         # The $(ls ...) shell expansion is done in the Github environment;
         # the value of TOGA_INSTALL_COMMAND will be a literal string without any shell expansions to perform
-        TOGA_INSTALL_COMMAND="python -m pip install ../$(ls dist/toga_core-*.whl)[dev] ../$(ls dist/toga_dummy-*.whl)" \
-          tox -e py-cov
-        tox -qe coverage$(tr -dc "0-9" <<< "${{ matrix.python-version }}")
-        mv core/.coverage core/.coverage.${{ matrix.platform }}.${{ matrix.python-version }}
+        TOGA_INSTALL_COMMAND="python -m pip install ../$(ls dist/toga_core-*.whl)[dev] ../$(ls dist/toga_dummy-*.whl) ../$(ls dist/travertino-*.whl)"
+        tox -e py-cov${{ matrix.env_suffix }}
+        tox -qe coverage$(tr -dc "0-9" <<< "${{ matrix.python-version }}"){{ matrix.env_suffix }}
+        mv ${{ matrix.package }}/.coverage ${{ matrix.package }}/.coverage.${{ matrix.platform }}.${{ matrix.python-version }}
 
     - name: Store Coverage Data
       uses: actions/upload-artifact@v4.5.0
       with:
-        name: core-coverage-data-${{ matrix.platform }}-${{ matrix.python-version }}
-        path: "core/.coverage.*"
+        name: ${{ matrix.package }}-coverage-data-${{ matrix.platform }}-${{ matrix.python-version }}
+        path: "$${{ matrix.coverage }}/.coverage.*"
         if-no-files-found: error
         include-hidden-files: true
 
@@ -159,15 +165,21 @@ jobs:
         with:
           python-version: "3.13"
 
-      - name: Get Packages
+      - name: Get Core Package
         uses: actions/download-artifact@v4.1.8
         with:
           name: Packages-toga-core
           path: dist
 
+      - name: Get Travertino Package
+        uses: actions/download-artifact@v4.1.8
+        with:
+          name: Packages-toga-travertino
+          path: dist
+
       - name: Test
         run: |
-          pip install dist/toga_core-*.whl
+          pip install dist/toga_core-*.whl dist/travertino-*.whl
           site_packages=$(python -c '
           import sys
           print([path for path in sys.path if "site-packages" in path][0])
@@ -176,17 +188,21 @@ jobs:
           cd core
           export MICROPYPATH="$site_packages:.frozen"
 
-          echo "Stable Travertino"
           ${{ steps.micropython.outputs.executable }} micropython_check.py
 
-          echo "Development Travertino"
-          pip install git+https://github.com/beeware/travertino
-          ${{ steps.micropython.outputs.executable }} micropython_check.py
-
-  core-coverage:
-    name: Coverage
+  core-and-travertino-coverage:
+    name: "Coverage: ${{ matrix.package }}"
     needs: core
     runs-on: ubuntu-latest
+    matrix:
+      package: ["core", "travertino"]
+      include:
+        - package: "core"
+          env_suffix: ""
+          fail: "-fail"
+        - package: "travertino"
+          env_suffix: "-trav"
+          fail: ""
     steps:
     - name: Checkout
       uses: actions/checkout@v4.2.2
@@ -210,19 +226,19 @@ jobs:
     - name: Retrieve Coverage Data
       uses: actions/download-artifact@v4.1.8
       with:
-        pattern: core-coverage-data-*
-        path: core
+        pattern: ${{ matrix.package }}-coverage-data-*
+        path: ${{ matrix.package }}
         merge-multiple: true
 
     - name: Generate Coverage Report
-      run: tox -e coverage-html-fail-platform
+      run: tox -e coverage-html${{ matrix.fail }}platform${{ matrix.env_suffix }}
 
     - name: Upload HTML Coverage Report
       uses: actions/upload-artifact@v4.5.0
       if: failure()
       with:
         name: html-coverage-report
-        path: core/htmlcov
+        path: ${{ matrix.package }}/htmlcov
 
   testbed:
     name: Testbed
