+++ tox.ini
@@ -9,13 +9,13 @@ extend-ignore =
 [tox]
 envlist = py{39,310,311,312,313}-cov,coverage
 labels =
-    test = py-cov,coverage
-    test39 = py39-cov,coverage39
-    test310 = py310-cov,coverage310
-    test311 = py311-cov,coverage311
-    test312 = py312-cov,coverage312
-    test313 = py313-cov,coverage313
-    ci = towncrier-check,docs-lint,pre-commit,py{39,310,311,312,313}-cov,coverage-fail-platform
+    test = py-cov{,-trav},coverage{,-trav}
+    test39 = py39-cov{,-trav},coverage39{,-trav}
+    test310 = py310-cov{,-trav},coverage310{,-trav}
+    test311 = py311-cov{,-trav},coverage311{,-trav}
+    test312 = py312-cov{,-trav},coverage312{,-trav}
+    test313 = py313-cov{,-trav},coverage313{,-trav}
+    ci = towncrier-check,docs-lint,pre-commit,py313-cov{,-trav},coverage{,-trav}-fail-platform
 skip_missing_interpreters = True
 
 [testenv:pre-commit]
@@ -25,9 +25,11 @@ deps =
 commands = pre-commit run --all-files --show-diff-on-failure --color=always
 
 # The leading comma generates the "py" environment
-[testenv:py{,39,310,311,312,313}{,-cov}]
+[testenv:py{,39,310,311,312,313}{,-cov}{,-trav}]
 depends = pre-commit
-changedir = core
+changedir =
+    !trav: core
+    trav: travertino
 skip_install = True
 setenv =
     TOGA_BACKEND = toga_dummy
@@ -35,14 +37,17 @@ allowlist_externals =
     bash
 commands =
     # TOGA_INSTALL_COMMAND is set to a bash command by the CI workflow
-    # Install as editable so parallel runs don't clobber the build directory for each other
     {env:TOGA_INSTALL_COMMAND:python -m pip install {tox_root}{/}core[dev] {tox_root}{/}dummy {tox_root}{/}travertino}
     !cov: python -m pytest {posargs:-vv --color yes}
     cov : python -m coverage run -m pytest {posargs:-vv --color yes}
 
-[testenv:coverage{,39,310,311,312,313}{,-html}{,-keep}{,-fail}{,-platform}]
-depends = pre-commit,py{,39,310,311,312,313}{,-cov}
-changedir = core
+[testenv:coverage{,39,310,311,312,313}{,-trav}{,-html}{,-keep}{,-fail}{,-platform}]
+depends =
+    !trav: pre-commit,py{,39,310,311,312,313}{,-cov}
+    trav: pre-commit,py{,39,310,311,312,313}{,-trav}{,-cov}
+changedir =
+    !trav: core
+    trav: travertino
 skip_install = True
 # by default, coverage should run on oldest supported Python for testing platform coverage.
 # however, coverage for a particular Python version should match the version used for pytest.
@@ -53,10 +58,14 @@ base_python =
     coverage312: py312
     coverage313: py313
 deps =
+    # Installing core[dev] even when just testing Travertino ensures the same versions
+    # of testing tools.
     {tox_root}{/}core[dev]
 setenv =
     keep: COMBINE_KEEP = --keep
     fail: REPORT_FAIL_COND = --fail-under=100
+    # Even when run with "fail", cut Travertino some slack.
+    trav: REPORT_FAIL_COND =
     CORE_RCFILE = --rcfile {tox_root}{/}core{/}pyproject.toml
     PROJECT_RCFILE = --rcfile {tox_root}{/}pyproject.toml
     # disable conditional coverage exclusions for Python version
@@ -90,6 +99,7 @@ suicide_timeout = 1
 deps =
     # editable install so docstrings can be updated for 'all' and 'live'
     -e {tox_root}{/}core[docs]
+    {tox_root}{/}travertino
 passenv =
     # On macOS M1, you need to manually set the location of the PyEnchant
     # library:
