+++ travertino/pyproject.toml
@@ -69,3 +69,6 @@ source = [
     "src/travertino",
     "**/travertino",
 ]
+
+[tool.pytest.ini_options]
+asyncio_default_fixture_loop_scope = "function"
