+++ travertino/tests/test_node.py
@@ -2,8 +2,8 @@ from unittest.mock import Mock
 from warnings import catch_warnings, filterwarnings
 
 import pytest
-from tests.travertino.utils import mock_attr, prep_style_class
 
+from tests.utils import mock_attr, prep_style_class
 from travertino.declaration import BaseStyle, Choices, validated_property
 from travertino.layout import BaseBox, Viewport
 from travertino.node import Node
